requirements.txt
numpy
tqdm
Pillow
imgaug
imagecorruptions
